6334 | 120 files | M DumpSource2/schemas/client/CDOTA_Ability_Beastmaster_PrimalRoar.h,  M DumpSource2/schemas/client/CDOTA_Ability_Grimstroke_DarkArtistry.h,  M DumpSource2/schemas/client/CDOTA_Ability_Marci_Companion_Run.h,  M DumpSource2/schemas/client/CDOTA_Ability_Snapfire_Scatterblast.h,  M DumpSource2/schemas/client/CDOTA_Ability_VoidSpirit_AstralStep.h,  M DumpSource2/schemas/client/CDOTA_AttackRecord.h,  M DumpSource2/schemas/client/CDOTA_Buff.h,  M DumpSource2/schemas/client/CDOTA_Modifier_Beastmaster_WildAxes.h,  M DumpSource2/schemas/client/CDOTA_Modifier_Bloodseeker_Bloodbath_Thinker.h,  M DumpSource2/schemas/client/CDOTA_Modifier_DarkSeer_Surge.h,  M DumpSource2/schemas/client/CDOTA_Modifier_Dawnbreaker_Luminosity_Attack_Buff.h,  M DumpSource2/schemas/client/CDOTA_Modifier_DoomBringer_ScorchedEarthEffectAura.h,  M DumpSource2/schemas/client/CDOTA_Modifier_DrowRanger_FrostArrows_Hypothermia.h,  M DumpSource2/schemas/client/CDOTA_Modifier_EmberSpirit_FireRemnant_RemnantTracker.h,  M DumpSource2/schemas/client/CDOTA_
https://steamdb.info/patchnotes/17492773/

+++ DumpSource2/schemas/client/CDOTA_Ability_Beastmaster_PrimalRoar.h
@@ -3,7 +3,7 @@ class CDOTA_Ability_Beastmaster_PrimalRoar
 	float32 duration;
 	float32 slow_duration;
 	int32 side_damage;
-	int32 damage_radius;
+	float32 damage_radius;
 	int32 path_width;
 	float32 push_duration;
 	float32 push_distance;
