+++ DumpSource2/schemas/client/CDOTA_Modifier_Beastmaster_WildAxes.h
@@ -11,9 +11,9 @@ class CDOTA_Modifier_Beastmaster_WildAxes
 	CUtlVector< CHandle< C_BaseEntity > > m_hAxes;
 	ParticleIndex_t[2] m_nAxeFXIndex;
 	CUtlVector< CHandle< C_BaseEntity > >[2] m_hHitList;
-	int32 radius;
-	int32 spread;
-	int32 axe_damage;
+	float32 radius;
+	float32 spread;
+	float32 axe_damage;
 	float32 duration;
 	float32 min_throw_duration;
 	float32 max_throw_duration;
