+++ DumpSource2/schemas/client/CDOTA_Modifier_DarkSeer_Surge.h
@@ -1,7 +1,7 @@
 class CDOTA_Modifier_DarkSeer_Surge
 {
 	int32 speed_boost;
-	int32 trail_radius;
+	float32 trail_radius;
 	float32 trail_duration;
 	Vector m_vLastTrailThinkerLocation;
 	bool m_bTrailStarted;
