+++ DumpSource2/schemas/client/CDOTA_Modifier_Dawnbreaker_Luminosity_Attack_Buff.h
@@ -3,7 +3,7 @@ class CDOTA_Modifier_Dawnbreaker_Luminosity_Attack_Buff
 	bool m_bIsBuffedAttack;
 	int32 heal_pct;
 	int32 bonus_damage;
-	int32 heal_radius;
+	float32 heal_radius;
 	int32 heal_from_creeps;
 	int32 allied_healing_pct;
 	bool triggered_by_celestial_hammer;
