+++ DumpSource2/schemas/client/CDOTA_Modifier_DrowRanger_FrostArrows_Hypothermia.h
@@ -1,6 +1,6 @@
 class CDOTA_Modifier_DrowRanger_FrostArrows_Hypothermia
 {
 	int32 shard_regen_reduction_pct_per_stack;
-	int32 shard_burst_radius;
+	float32 shard_burst_radius;
 	ParticleIndex_t m_nFXStackIndex;
 };
