+++ DumpSource2/schemas/client/CDOTA_Modifier_Marci_Grapple_VictimMotion.h
@@ -2,5 +2,5 @@ class CDOTA_Modifier_Marci_Grapple_VictimMotion
 {
 	int32 impact_damage;
 	float32 debuff_duration;
-	int32 landing_radius;
+	float32 landing_radius;
 };
