+++ DumpSource2/schemas/client/CDOTA_Modifier_StormSpirit_StaticRemnantThinker.h
@@ -1,8 +1,8 @@
 class CDOTA_Modifier_StormSpirit_StaticRemnantThinker
 {
-	int32 static_remnant_radius;
-	int32 static_remnant_damage_radius;
-	int32 static_remnant_damage;
+	float32 static_remnant_radius;
+	float32 static_remnant_damage_radius;
+	float32 static_remnant_damage;
 	float32 static_remnant_delay;
 	float32 static_remnant_travel_speed;
 	int32 m_iSearchRadius;
