+++ DumpSource2/schemas/client/CDOTA_Modifier_TemplarAssassin_PsiBlades.h
@@ -1,11 +1,11 @@
 class CDOTA_Modifier_TemplarAssassin_PsiBlades
 {
-	int32 bonus_attack_range;
+	float32 bonus_attack_range;
 	float32 attack_spill_range;
-	int32 attack_spill_width;
+	float32 attack_spill_width;
 	int32 attack_spill_pct;
 	int32 attack_spill_penalty;
-	int32 additional_bonus_attack_range;
-	int32 additional_attack_spill_range;
+	float32 additional_bonus_attack_range;
+	float32 additional_attack_spill_range;
 	float32 spill_slow_duration;
 };
