+++ DumpSource2/schemas/client/CDOTA_Modifier_VoidSpirit_ResonantPulse_Ring.h
@@ -5,7 +5,7 @@ class CDOTA_Modifier_VoidSpirit_ResonantPulse_Ring
 	GameTime_t m_fLastThink;
 	ParticleIndex_t m_nFXIndex;
 	CUtlVector< CHandle< C_BaseEntity > > m_EntitiesHit;
-	int32 speed;
-	int32 radius;
-	int32 damage;
+	float32 speed;
+	float32 radius;
+	float32 damage;
 };
