+++ DumpSource2/schemas/client/C_DOTA_Ability_Beastmaster_WildAxes.h
@@ -1,4 +1,4 @@
 class C_DOTA_Ability_Beastmaster_WildAxes
 {
-	int32 axe_damage;
+	float32 axe_damage;
 };
