+++ DumpSource2/schemas/client/C_DOTA_Ability_Dazzle_ShadowWave.h
@@ -1,8 +1,8 @@
 class C_DOTA_Ability_Dazzle_ShadowWave
 {
 	CUtlVector< CHandle< C_BaseEntity > > m_hHitEntities;
-	int32 bounce_radius;
-	int32 damage_radius;
+	float32 bounce_radius;
+	float32 damage_radius;
 	int32 damage;
 	int32 max_targets;
 	float32 scepter_heal_pct;
