+++ DumpSource2/schemas/client/C_DOTA_Ability_DeathProphet_CarrionSwarm.h
@@ -1,7 +1,7 @@
 class C_DOTA_Ability_DeathProphet_CarrionSwarm
 {
-	int32 start_radius;
-	int32 end_radius;
+	float32 start_radius;
+	float32 end_radius;
 	GameTime_t m_fStartTime;
 	float32 m_fTotalTime;
 	int32 m_nProjectileHandle;
