+++ DumpSource2/schemas/client/C_DOTA_Ability_DrowRanger_WaveOfSilence.h
@@ -1,6 +1,6 @@
 class C_DOTA_Ability_DrowRanger_WaveOfSilence
 {
-	int32 wave_width;
+	float32 wave_width;
 	float32 wave_speed;
 	float32 silence_duration;
 	float32 knockback_distance_max;
