+++ DumpSource2/schemas/client/C_DOTA_Ability_Hoodwink_Bushwhack.h
@@ -1,6 +1,6 @@
 class C_DOTA_Ability_Hoodwink_Bushwhack
 {
-	int32 trap_radius;
+	float32 trap_radius;
 	int32 m_nProjectileHandle;
 	ParticleIndex_t m_nFXIndex;
 };
