+++ DumpSource2/schemas/client/C_DOTA_Ability_Muerta_DeadShot.h
@@ -1,5 +1,5 @@
 class C_DOTA_Ability_Muerta_DeadShot
 {
-	int32 radius;
-	int32 speed;
+	float32 radius;
+	float32 speed;
 };
