+++ DumpSource2/schemas/client/C_DOTA_Ability_Nyx_Assassin_Impale.h
@@ -1,9 +1,9 @@
 class C_DOTA_Ability_Nyx_Assassin_Impale
 {
 	CUtlVector< CHandle< C_BaseEntity > > hAlreadyHitList;
-	int32 width;
+	float32 width;
 	float32 duration;
-	int32 length;
-	int32 speed;
+	float32 length;
+	float32 speed;
 	Vector vOriginalCast;
 };
