+++ DumpSource2/schemas/server/CDOTA_Ability_DrowRanger_WaveOfSilence.h
@@ -2,7 +2,7 @@ class CDOTA_Ability_DrowRanger_WaveOfSilence
 {
 	Vector m_vStartPosition;
 	Vector m_vEndPosition;
-	int32 wave_width;
+	float32 wave_width;
 	float32 wave_speed;
 	float32 silence_duration;
 	float32 knockback_distance_max;
