+++ DumpSource2/schemas/server/CDOTA_Ability_Hoodwink_Bushwhack.h
@@ -1,6 +1,6 @@
 class CDOTA_Ability_Hoodwink_Bushwhack
 {
-	int32 trap_radius;
+	float32 trap_radius;
 	int32 m_nProjectileHandle;
 	ParticleIndex_t m_nFXIndex;
 };
