+++ DumpSource2/schemas/server/CDOTA_Ability_Muerta_DeadShot.h
@@ -1,7 +1,7 @@
 class CDOTA_Ability_Muerta_DeadShot
 {
-	int32 radius;
-	int32 speed;
+	float32 radius;
+	float32 speed;
 	Vector m_vTargetPos;
 	Vector m_vEndpoint;
 	CHandle< CBaseEntity > m_hTreeTarget;
