+++ DumpSource2/schemas/server/CDOTA_Ability_Nevermore_Requiem.h
@@ -1,7 +1,7 @@
 class CDOTA_Ability_Nevermore_Requiem
 {
-	int32 requiem_line_width_start;
-	int32 requiem_line_width_end;
+	float32 requiem_line_width_start;
+	float32 requiem_line_width_end;
 	int32 m_nCachedSouls;
 	ParticleIndex_t m_nFXIndex;
 	int32 m_nKilleater_nLines;
