+++ DumpSource2/schemas/server/CDOTA_Ability_Nyx_Assassin_Impale.h
@@ -1,9 +1,9 @@
 class CDOTA_Ability_Nyx_Assassin_Impale
 {
 	CUtlVector< CHandle< CBaseEntity > > hAlreadyHitList;
-	int32 width;
+	float32 width;
 	float32 duration;
-	int32 length;
-	int32 speed;
+	float32 length;
+	float32 speed;
 	Vector vOriginalCast;
 };
