+++ DumpSource2/schemas/server/CDOTA_Ability_Shredder_WhirlingDeath.h
@@ -1,6 +1,6 @@
 class CDOTA_Ability_Shredder_WhirlingDeath
 {
-	int32 whirling_radius;
+	float32 whirling_radius;
 	int32 whirling_damage;
 	float32 whirling_tick;
 	float32 duration;
