+++ DumpSource2/schemas/server/CDOTA_Ability_Skywrath_Mage_Concussive_Shot.h
@@ -1,9 +1,9 @@
 class CDOTA_Ability_Skywrath_Mage_Concussive_Shot
 {
-	int32 speed;
-	int32 slow_radius;
-	int32 damage;
-	int32 shot_vision;
+	float32 speed;
+	float32 slow_radius;
+	float32 damage;
+	float32 shot_vision;
 	float32 slow_duration;
 	float32 vision_duration;
 };
