+++ DumpSource2/schemas/server/CDOTA_Ability_Snapfire_Scatterblast.h
@@ -4,5 +4,5 @@ class CDOTA_Ability_Snapfire_Scatterblast
 	float32 debuff_duration;
 	int32 point_blank_range;
 	float32 point_blank_dmg_bonus_pct;
-	int32 blast_width_end;
+	float32 blast_width_end;
 };
