+++ DumpSource2/schemas/server/CDOTA_Ability_Tusk_IceShards.h
@@ -1,13 +1,13 @@
 class CDOTA_Ability_Tusk_IceShards
 {
 	int32 m_iProjectile;
-	int32 shard_width;
-	int32 shard_damage;
+	float32 shard_width;
+	float32 shard_damage;
 	int32 shard_count;
 	float32 shard_speed;
 	float32 shard_duration;
 	float32 shard_angle_step;
-	int32 shard_distance;
+	float32 shard_distance;
 	Vector m_vSpawnOrigin;
 	Vector m_vDirection;
 	CUtlVector< CHandle< CBaseEntity > > m_vecShards;
