+++ DumpSource2/schemas/server/CDOTA_Ability_VoidSpirit_AstralStep.h
@@ -3,9 +3,9 @@ class CDOTA_Ability_VoidSpirit_AstralStep
 	Vector m_vStartPos;
 	Vector m_vDestination;
 	Vector m_vDirection;
-	int32 radius;
-	int32 min_travel_distance;
-	int32 max_travel_distance;
+	float32 radius;
+	float32 min_travel_distance;
+	float32 max_travel_distance;
 	float32 debuff_duration;
 	float32 pop_damage_delay;
 };
