+++ DumpSource2/schemas/server/CDOTA_Modifier_EmberSpirit_FireRemnant_RemnantTracker.h
@@ -2,5 +2,5 @@ class CDOTA_Modifier_EmberSpirit_FireRemnant_RemnantTracker
 {
 	CUtlVector< CHandle< CBaseEntity > > m_vActiveRemnants;
 	CUtlVector< CHandle< CBaseEntity > > m_vTrackingDelayedRemnants;
-	int32 shard_charge_radius;
+	float32 shard_charge_radius;
 };
