+++ DumpSource2/schemas/server/CDOTA_Modifier_Invoker_SunStrike.h
@@ -1,8 +1,8 @@
 class CDOTA_Modifier_Invoker_SunStrike
 {
-	int32 area_of_effect;
+	float32 area_of_effect;
 	float32 damage;
-	int32 vision_distance;
+	float32 vision_distance;
 	float32 vision_duration;
 	float32 spell_amp;
 	float32 spell_lifesteal;
