+++ DumpSource2/schemas/server/CDOTA_Modifier_Item_Battlefury.h
@@ -5,9 +5,9 @@ class CDOTA_Modifier_Item_Battlefury
 	float32 bonus_mana_regen;
 	int32 cleave_damage_percent;
 	int32 cleave_damage_percent_creep;
-	int32 cleave_starting_width;
-	int32 cleave_ending_width;
-	int32 cleave_distance;
+	float32 cleave_starting_width;
+	float32 cleave_ending_width;
+	float32 cleave_distance;
 	int32 upgraded_cleave_bonus;
 	int32 upgraded_damage_bonus;
 	int32 quelling_bonus;
