+++ DumpSource2/schemas/server/CDOTA_Modifier_Jakiro_IcePath_Thinker.h
@@ -1,7 +1,7 @@
 class CDOTA_Modifier_Jakiro_IcePath_Thinker
 {
 	float32 path_delay;
-	int32 path_radius;
+	float32 path_radius;
 	int32 m_nDamage;
 	float32 m_flRadius;
 	float32 stun_duration;
