+++ DumpSource2/schemas/server/CDOTA_Modifier_Marci_Grapple_VictimMotion.h
@@ -2,7 +2,7 @@ class CDOTA_Modifier_Marci_Grapple_VictimMotion
 {
 	int32 impact_damage;
 	float32 debuff_duration;
-	int32 landing_radius;
+	float32 landing_radius;
 	float32 air_duration;
 	int32 air_height;
 	int32 travel_distance;
