+++ DumpSource2/schemas/server/CDOTA_Modifier_Techies_ReactiveTazer.h
@@ -1,7 +1,7 @@
 class CDOTA_Modifier_Techies_ReactiveTazer
 {
 	int32 bonus_ms;
-	int32 stun_radius;
+	float32 stun_radius;
 	float32 stun_duration;
 	int32 damage_percent;
 	int32 m_nDamageRemaining;
